+++ packages/ui/portal.manifest.mjs
@@ -1,6 +1,6 @@
 export default {
   id: 'ui',
-  version: '0.4.5',
+  version: '0.4.6',
   source: '@nuxt-customer-portal/ui',
   dependsOn: ['core']
 }