+++ packages/kit/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @nuxt-customer-portal/kit
 
+## 0.4.6
+
 ## 0.4.5
 
 ## 0.4.4