+++ packages/preset/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @nuxt-customer-portal/preset
 
+## 0.4.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @nuxt-customer-portal/core@0.4.6
+  - @nuxt-customer-portal/administration@0.4.6
+  - @nuxt-customer-portal/authentication@0.4.6
+  - @nuxt-customer-portal/clients@0.4.6
+  - @nuxt-customer-portal/organizations@0.4.6
+  - @nuxt-customer-portal/ui@0.4.6
+
 ## 0.4.5
 
 ### Patch Changes