+++ packages/service-requests/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nuxt-customer-portal/service-requests
 
+## 0.4.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @nuxt-customer-portal/core@0.4.6
+  - @nuxt-customer-portal/clients@0.4.6
+
 ## 0.4.5
 
 ### Patch Changes