+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nuxt-customer-portal/ui
 
+## 0.4.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @nuxt-customer-portal/core@0.4.6
+
 ## 0.4.5
 
 ### Patch Changes